chore: creates HexagonCell.
