+++ src/components/board/index.tsx
@@ -1,0 +1,3 @@
+export { default as Board } from './board'
+export { default as HexagonGrid } from './hexagonGrid'
+export { default as HexagonCell } from './hexagonCell'
