+++ src/components/board/styles.tsx
