--- src/components/hexagon/index.tsx
@@ -1,2 +1,0 @@
-export { default as Hexagon } from './hexagon'
-export { default as HexagonGrid } from './hexagonGrid'
