+++ src/game/Board.ts
@@ -4,7 +4,7 @@ import { Obstacle } from './Obstacle'
 export const GRID_SIZE = 53
 
 export interface Cell {
-  obstacle: Obstacle | undefined
+  obstacle?: Obstacle
   position: number
-  player: Player | undefined
+  player?: Player
 }
